Reworks path to allow code refactor
Instead of declaring one type and pushFunction for each Path inside the
starting page, we can now use a single function to push any section to
the appHistory.

Note that this is intended to be used on the redesign and currently do
not shows a lot of benefit on the master branch.